+++ src/views/UserEdit/UserForm.js
@@ -26,7 +26,10 @@ import {
 import { EditCustomFieldsRecord } from '@folio/stripes/smart-components';
 import stripesFinalForm from '@folio/stripes/final-form';
 
-import { USER_TYPES } from '../../constants';
+import {
+  CUSTOM_FIELDS_SCOPE,
+  USER_TYPES,
+} from '../../constants';
 import {
   EditUserInfo,
   EditExtendedInfo,
@@ -445,6 +448,7 @@ class UserForm extends React.Component {
                     finalFormCustomFieldsValues={form.getState().values.customFields}
                     fieldComponent={Field}
                     changeFinalFormField={form.change}
+                    scope={CUSTOM_FIELDS_SCOPE}
                   />
                   {initialValues.id &&
                     <div>